+++ src/Core.ts
@@ -254,7 +254,7 @@ export default class Core {
    * @param {string} resourceID id of the Resource
    * @returns {Promise<Resource>} resolves to the Resource which should be loaded
    */
-  resource(relation: string, resourceID): Promise<Resource> {
+  resource(relation: string, resourceID, additionalTemplateParams: any = {}): Promise<Resource> {
     return Promise.resolve()
     .then(() => {
       if (!relation) {
@@ -270,7 +270,8 @@ export default class Core {
       return this.follow(this[relationsSymbol][relation].relation)
     })
     .then((request) => {
-      request.withTemplateParameters({ [this[relationsSymbol][relation].id]: resourceID });
+      const params = Object.assign({}, additionalTemplateParams, { [this[relationsSymbol][relation].id]: resourceID });
+      request.withTemplateParameters(params);
       return get(this[environmentSymbol], request);
     })
     .then(([res, traversal]) =>