+++ test/publicAPI/PublicAPI.test.js
@@ -20,6 +20,8 @@ const EntryList = require('../../lib/resources/publicAPI/EntryList').default;
 const EntryResource = require('../../lib/resources/publicAPI/EntryResource').default;
 const PublicAssetList = require('../../lib/resources/publicAPI/PublicAssetList').default;
 const PublicAssetResource = require('../../lib/resources/publicAPI/PublicAssetResource').default;
+const DMAssetList = require('../../lib/resources/publicAPI/DMAssetList').default;
+const DMAssetResource = require('../../lib/resources/publicAPI/DMAssetResource').default;
 
 const environmentSymbol = Symbol.for('environment');
 const resourceSymbol = Symbol.for('resource');
@@ -923,7 +925,16 @@ describe('PublicAPI', () => {
     stub.onFirstCall().returns(resolver('public-dm-root.json'));
     stub.onSecondCall().returns(resolver('best-file.json'));
 
-    return api.getFileUrl('id')
+    return api.getFileUrl('e766d956-6f43-49fa-8f30-023e4cd29779')
+    .should.eventually.be.equal('https://cdn2.entrecode.de/files/01bd8e08/J2DJfjfEVby3KcxGNrJyFdEz_512.png')
+    .notify(() => stub.restore());
+  });
+  it('should get best file assetNeue', () => {
+    const stub = sinon.stub(helper, 'get');
+    stub.onFirstCall().returns(resolver('public-dm-root.json'));
+    stub.onSecondCall().returns(resolver('best-file.json'));
+
+    return api.getFileUrl('notUUID')
     .should.eventually.be.equal('https://cdn2.entrecode.de/files/01bd8e08/J2DJfjfEVby3KcxGNrJyFdEz_512.png')
     .notify(() => stub.restore());
   });
@@ -936,7 +947,7 @@ describe('PublicAPI', () => {
     stub.onFirstCall().returns(resolver('public-dm-root.json'));
     stub.onSecondCall().returns(resolver('best-file.json'));
 
-    return api.getImageUrl('id')
+    return api.getImageUrl('e766d956-6f43-49fa-8f30-023e4cd29779')
     .should.eventually.be.equal('https://cdn2.entrecode.de/files/01bd8e08/J2DJfjfEVby3KcxGNrJyFdEz_512.png')
     .notify(() => stub.restore());
   });
@@ -945,7 +956,7 @@ describe('PublicAPI', () => {
     stub.onFirstCall().returns(resolver('public-dm-root.json'));
     stub.onSecondCall().returns(resolver('best-file.json'));
 
-    return api.getImageUrl('id', 2)
+    return api.getImageUrl('e766d956-6f43-49fa-8f30-023e4cd29779', 2)
     .should.eventually.be.equal('https://cdn2.entrecode.de/files/01bd8e08/J2DJfjfEVby3KcxGNrJyFdEz_512.png')
     .notify(() => stub.restore());
   });
@@ -958,7 +969,16 @@ describe('PublicAPI', () => {
     stub.onFirstCall().returns(resolver('public-dm-root.json'));
     stub.onSecondCall().returns(resolver('best-file.json'));
 
-    return api.getImageThumbUrl('id')
+    return api.getImageThumbUrl('e766d956-6f43-49fa-8f30-023e4cd29779')
+    .should.eventually.be.equal('https://cdn2.entrecode.de/files/01bd8e08/J2DJfjfEVby3KcxGNrJyFdEz_512.png')
+    .notify(() => stub.restore());
+  });
+  it('should get best thumb assetNeue', () => {
+    const stub = sinon.stub(helper, 'get');
+    stub.onFirstCall().returns(resolver('public-dm-root.json'));
+    stub.onSecondCall().returns(resolver('best-file.json'));
+
+    return api.getImageThumbUrl('notUUID')
     .should.eventually.be.equal('https://cdn2.entrecode.de/files/01bd8e08/J2DJfjfEVby3KcxGNrJyFdEz_512.png')
     .notify(() => stub.restore());
   });
@@ -967,7 +987,7 @@ describe('PublicAPI', () => {
     stub.onFirstCall().returns(resolver('public-dm-root.json'));
     stub.onSecondCall().returns(resolver('best-file.json'));
 
-    return api.getImageThumbUrl('id', 2)
+    return api.getImageThumbUrl('e766d956-6f43-49fa-8f30-023e4cd29779', 2)
     .should.eventually.be.equal('https://cdn2.entrecode.de/files/01bd8e08/J2DJfjfEVby3KcxGNrJyFdEz_512.png')
     .notify(() => stub.restore());
   });
@@ -992,5 +1012,241 @@ describe('PublicAPI', () => {
   });
   it('should be rejected on undefined email', () => {
     return api.changeEmail().should.be.rejectedWith('email must be defined');
+  });
+
+  it('should load asset list neue', () => {
+    const stub = sinon.stub(helper, 'get');
+    stub.onFirstCall().returns(resolver('public-dm-root.json'));
+    stub.onSecondCall().returns(resolver('dm-asset-list.json'));
+
+    return api.dmAssetList('test1')
+    .then((list) => {
+      list.should.be.instanceof(DMAssetList);
+      stub.restore();
+    })
+    .catch((err) => {
+      stub.restore();
+      throw err;
+    });
+  });
+  it('should be rejected on undefined assetGroupID', () => {
+    return api.dmAssetList().should.be.rejectedWith('assetGroupID must be defined');
+  });
+  it('should be rejected on unknown assetGroupID', () => {
+    return api.dmAssetList('unknown').should.be.rejectedWith('assetGroup not found');
+  });
+  it('should throw on asset list neue filtered with assetID', () => {
+    return api.dmAssetList('test1', { assetID: 'id' })
+    .should.be.rejectedWith('Cannot filter assetList only by assetID. Use PublicAPI#dmAsset() instead');
+  });
+  it('should load asset resource neue', () => {
+    const stub = sinon.stub(helper, 'get');
+    stub.onFirstCall().returns(resolver('public-dm-root.json'));
+    stub.onSecondCall().returns(resolver('dm-asset-single.json'));
+
+    return api.dmAsset('test1', 'id')
+    .then((model) => {
+      model.should.be.instanceof(DMAssetResource);
+      stub.restore();
+    })
+    .catch((err) => {
+      stub.restore();
+      throw err;
+    });
+  });
+  it('should be rejected on undefined assetID', () => {
+    return api.dmAsset('test1').should.be.rejectedWith('assetID must be defined');
+  });
+  it('should be rejected on undefined assetGroupID', () => {
+    return api.dmAsset().should.be.rejectedWith('assetGroupID must be defined');
+  });
+  it('should be rejected on unknown assetGroupID', () => {
+    return api.dmAsset('unknown', 'id').should.be.rejectedWith('assetGroup not found');
+  });
+
+  describe('dmAssets', () => {
+    it('should create dmAssets, path #1', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+      const stubSuperagentPost = sinon.stub(helper, 'superagentPost');
+      stubSuperagentPost.returns(resolver('dm-asset-create.json'));
+
+      return api.createDMAssets('test1', `${__dirname}/../mocks/test.png`)
+      .then((response) => {
+        response.should.be.instanceof(DMAssetList);
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+        throw err;
+      });
+    });
+    it('should create dmAssets, path #2', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+      const stubSuperagentPost = sinon.stub(helper, 'superagentPost');
+      stubSuperagentPost.returns(resolver('dm-asset-create.json'));
+
+      return api.createDMAssets('test1', `${__dirname}/../mocks/test.png`, { fileName: 'name' })
+      .then((response) => {
+        response.should.be.instanceof(DMAssetList);
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+        throw err;
+      });
+    });
+    it('should create dmAssets, path array', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+      const stubSuperagentPost = sinon.stub(helper, 'superagentPost');
+      stubSuperagentPost.returns(resolver('dm-asset-create.json'));
+
+      return api.createDMAssets('test1', [`${__dirname}/../mocks/test.png`, `${__dirname}/../mocks/test.png`])
+      .then((response) => {
+        response.should.be.instanceof(DMAssetList);
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+        throw err;
+      });
+    });
+    it('should create dmAssets, buffer, title and tags', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+      const stubSuperagentPost = sinon.stub(helper, 'superagentPost');
+      stubSuperagentPost.returns(resolver('dm-asset-create.json'));
+
+      return new Promise((resolve, reject) => {
+        fs.readFile(`${__dirname}/../mocks/test.png`, (err, file) => {
+          if (err) {
+            return reject(err);
+          }
+          return resolve(file);
+        });
+      })
+      .then(file => api.createDMAssets('test1', [file, file], {
+        fileName: ['test.png', 'test.png'],
+        title: 'hello',
+        tags: ['helloTag'],
+      }))
+      .then((response) => {
+        response.should.be.instanceof(DMAssetList);
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+        throw err;
+      });
+    });
+    it('should be rejected on create dmAssets with buffer and no file name #1', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+
+      return api.createDMAssets('test1', [new Buffer([])])
+      .then(() => {
+        throw new Error('Unexpectedly resolved');
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        if (err.message === 'Unexpectedly resolved') {
+          throw err;
+        }
+        err.message.should.be.equal('When using buffer file input you must provide options.fileName.');
+      });
+    });
+    it('should be rejected on create dmAssets with buffer and no file name #2', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+
+      return api.createDMAssets('test1', [new Buffer([])], { fileName: 'string' })
+      .then(() => {
+        throw new Error('Unexpectedly resolved');
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        if (err.message === 'Unexpectedly resolved') {
+          throw err;
+        }
+        err.message.should.be.equal('When using buffer file input you must provide options.fileName.');
+      });
+    });
+    it('should be rejected on create dmAssets with buffer and no file name #3', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+
+      return api.createDMAssets('test1', [new Buffer([])], { fileName: [] })
+      .then(() => {
+        throw new Error('Unexpectedly resolved');
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        if (err.message === 'Unexpectedly resolved') {
+          throw err;
+        }
+        err.message.should.be.equal('When using buffer file input you must provide options.fileName.');
+      });
+    });
+    it('should create dmAssets, FormData (mock), title and tags', () => {
+      mock.reset();
+      global.FormData = fdMock;
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/a/beefbeef/test1'));
+      const stubSuperagentPost = sinon.stub(helper, 'superagentPost');
+      stubSuperagentPost.returns(resolver('dm-asset-create.json'));
+
+      return api.createDMAssets('test1', new FormData(), { // eslint-disable-line no-undef
+        title: 'hello',
+        tags: ['whatwhat'],
+      })
+      .then((response) => {
+        response.should.be.instanceof(DMAssetList);
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+        global.FormData = undefined;
+      })
+      .catch((err) => {
+        stubGetUrl.restore();
+        stubSuperagentPost.restore();
+        global.FormData = undefined;
+        throw err;
+      });
+    });
+    it('should be rejected on create dmAssets with undefined assetGroupID', () => {
+      mock.reset();
+      return api.createDMAssets().should.be.rejectedWith('assetGroupID must be defined');
+    });
+    it('should be rejected on create dmAssets with unknown assetGroupID', () => {
+      mock.reset();
+      return api.createDMAssets('asdf', 'test').should.be.rejectedWith('assetGroup not found');
+    });
+    it('should be rejected on create dmAssets with undefined value', () => {
+      mock.reset();
+      return api.createDMAssets('test1').should.be.rejectedWith('Cannot create resource with undefined object.');
+    });
+    it('should be rejected on create dmAssets with unsupported value', () => {
+      mock.reset();
+      const stubGetUrl = sinon.stub(helper, 'getUrl');
+      stubGetUrl.returns(Promise.resolve('https://datamanager.entrecode.de/asset/beefbeef'));
+      return api.createDMAssets('test1', [[]]).should.be.rejectedWith('Cannot handle input.')
+      .notify(() => stubGetUrl.restore());
+    });
   });
 });